Added settingValueOfElement method
Added new method -

`settingValueOfElement(sValue, dDateTime, oInputElement) `
This callback function will be called before setting value in the input
field.
`sValue` is the DateTime string to be set in the Input field.
`dDateTime` is the Javascript Date Object of DateTime set in Input
Field.
`oInputElement` is the Input Element in which value is to be set.

This method is added primarily to make DateTimePicker work with
AngularJS.

See. [Angular
Demo](https://github.com/CuriousSolutions/DateTimePicker/blob/master/dem
o/AngularJS/BasicDatePicker-AngularJS.htm).