Added Format as a parameter in callbacks
Added Format as a parameter in `parseDateTimeString` and
`formatDateTimeString` callbacks to let it support different formats.
Previously, with this function parsing or formatting of only one format
was possible. Now based on format different code can be written for
parsing or formatting the date.

```
parseDateTimeString: function(sDateTime, sMode, sFormat, oInputElement)
{
}
```

```
formatDateTimeString: function(oDate, sMode, sFormat, oInputElement) {
}
```
Added exmaple in which date format of date string in input field is
different from the date format of date string to be displayed on
DateTimePicker, as requested in #125

Added exmaple for using different date formats for storing date in the
database and displaying on DateTimePicker, as requested in #122